+++ src/cosmos/CosmosEngine.ts
@@ -1,0 +1,380 @@
+import { decodeSignature, encodeSecp256k1Pubkey } from '@cosmjs/amino'
+import { encodePubkey, makeAuthInfoBytes } from '@cosmjs/proto-signing'
+import { StargateClient } from '@cosmjs/stargate'
+import { add, gt } from 'biggystring'
+import { SignDoc, TxBody, TxRaw } from 'cosmjs-types/cosmos/tx/v1beta1/tx'
+import {
+  EdgeCurrencyEngine,
+  EdgeCurrencyEngineOptions,
+  EdgeFetchFunction,
+  EdgeFreshAddress,
+  EdgeMemo,
+  EdgeSpendInfo,
+  EdgeTransaction,
+  EdgeWalletInfo,
+  InsufficientFundsError,
+  JsonObject,
+  NoAmountSpecifiedError
+} from 'edge-core-js/types'
+import { base16 } from 'rfc4648'
+
+import { CurrencyEngine } from '../common/CurrencyEngine'
+import { PluginEnvironment } from '../common/innerPlugin'
+import { upgradeMemos } from '../common/upgradeMemos'
+import { cleanTxLogs, getFetchCors } from '../common/utils'
+import { CosmosTools } from './CosmosTools'
+import {
+  asCosmosPrivateKeys,
+  asCosmosWalletOtherData,
+  asSafeCosmosWalletInfo,
+  asShapeshiftResponse,
+  CosmosNetworkInfo,
+  CosmosWalletOtherData,
+  SafeCosmosWalletInfo,
+  ShapeshiftTx
+} from './cosmosTypes'
+
+const ACCOUNT_POLL_MILLISECONDS = 5000
+const TRANSACTION_POLL_MILLISECONDS = 3000
+
+export class CosmosEngine extends CurrencyEngine<
+  CosmosTools,
+  SafeCosmosWalletInfo
+> {
+  networkInfo: CosmosNetworkInfo
+  accountNumber: number
+  sequence: number
+  fetchCors: EdgeFetchFunction
+  otherData!: CosmosWalletOtherData
+
+  constructor(
+    env: PluginEnvironment<CosmosNetworkInfo>,
+    tools: CosmosTools,
+    walletInfo: SafeCosmosWalletInfo,
+    opts: EdgeCurrencyEngineOptions
+  ) {
+    super(env, tools, walletInfo, opts)
+    this.networkInfo = env.networkInfo
+    this.accountNumber = 0
+    this.sequence = 0
+    this.fetchCors = getFetchCors(env.io)
+  }
+
+  setOtherData(raw: any): void {
+    this.otherData = asCosmosWalletOtherData(raw)
+  }
+
+  async getStargateClient(): Promise<StargateClient> {
+    if (this.tools.client == null) {
+      throw new Error('No StargateClient')
+    }
+    return this.tools.client
+  }
+
+  async queryBalance(): Promise<void> {
+    try {
+      const client = await this.getStargateClient()
+      const balances = await client.getAllBalances(
+        this.walletInfo.keys.bech32Address
+      )
+      const mainnetBal = balances.find(
+        bal => bal.denom === this.currencyInfo.currencyCode.toLowerCase()
+      )
+      this.updateBalance(
+        this.currencyInfo.currencyCode,
+        mainnetBal?.amount ?? '0'
+      )
+
+      const { accountNumber, sequence } = await client.getSequence(
+        this.walletInfo.keys.bech32Address
+      )
+      this.accountNumber = accountNumber
+      this.sequence = sequence
+    } catch (e) {
+      if (String(e).includes('does not exist on chain')) {
+        this.updateBalance(this.currencyInfo.currencyCode, '0')
+      } else {
+        this.log.warn('queryBalance error:', e)
+      }
+    }
+  }
+
+  async queryBlockheight(): Promise<void> {
+    try {
+      const client = await this.getStargateClient()
+      const blockheight = await client.getHeight()
+      if (blockheight > this.walletLocalData.blockHeight) {
+        this.walletLocalData.blockHeight = blockheight
+        this.walletLocalDataDirty = true
+        this.currencyEngineCallbacks.onBlockHeightChanged(
+          this.walletLocalData.blockHeight
+        )
+      }
+    } catch (e: any) {
+      this.error(`queryBlockheight Error `, e)
+    }
+  }
+
+  async queryTransactions(): Promise<void> {
+    let cursor: string | undefined
+    let firstTxid: string | undefined
+    let firstTxidIndex: number | undefined
+    try {
+      do {
+        const res = await this.fetchCors(
+          `https://api.${
+            this.networkInfo.shapeshiftApiName
+          }.shapeshift.com/api/v1/account/${
+            this.walletInfo.keys.bech32Address
+          }/txs?pageSize=100&cursor=${cursor ?? ''}`
+        )
+        const rawJson = await res.json()
+        const json = asShapeshiftResponse(rawJson)
+        cursor = json.cursor
+        for (const [i, tx] of json.txs.entries()) {
+          if (tx == null) continue
+          if (firstTxid == null) {
+            firstTxid = tx.txid
+            firstTxidIndex = i
+          }
+          if (tx.txid === this.otherData.newestTxid) {
+            break
+          }
+          this.processCosmosTransaction(tx)
+        }
+      } while (cursor != null)
+    } catch (e) {
+      this.log.warn('queryTransactions error:', e)
+      throw e
+    }
+
+    if (
+      firstTxid != null &&
+      firstTxidIndex !== this.otherData.newestTxidIndex
+    ) {
+      this.otherData.newestTxid = firstTxid
+      this.otherData.newestTxidIndex = firstTxidIndex
+      this.walletLocalDataDirty = true
+    }
+    this.tokenCheckTransactionsStatus[this.currencyInfo.currencyCode] = 1
+    this.updateOnAddressesChecked()
+
+    if (this.transactionsChangedArray.length > 0) {
+      this.currencyEngineCallbacks.onTransactionsChanged(
+        this.transactionsChangedArray
+      )
+      this.transactionsChangedArray = []
+    }
+  }
+
+  processCosmosTransaction(tx: ShapeshiftTx): void {
+    const { blockHeight, fee, memo, messages, timestamp, txid } = tx
+    const message = messages[0]
+    const { from, value } = message
+
+    const { amount: networkFee } = fee
+
+    const memos: EdgeMemo[] = []
+    if (memo != null) {
+      memos.push({
+        type: 'text',
+        value: memo
+      })
+    }
+
+    const isSend = from === this.walletInfo.keys.bech32Address
+    const ourReceiveAddresses: string[] = []
+
+    const currencyCode = value.denom.toUpperCase()
+    let nativeAmount = value.amount
+
+    if (isSend) {
+      nativeAmount = `-${add(nativeAmount, networkFee)}`
+    } else {
+      ourReceiveAddresses.push(this.walletInfo.keys.bech32Address)
+    }
+
+    const edgeTransaction: EdgeTransaction = {
+      blockHeight,
+      currencyCode,
+      date: timestamp,
+      isSend,
+      memos,
+      nativeAmount,
+      networkFee,
+      ourReceiveAddresses,
+      signedTx: '',
+      txid,
+      walletId: this.walletId
+    }
+    this.addTransaction(currencyCode, edgeTransaction)
+  }
+
+  // // ****************************************************************************
+  // // Public methods
+  // // ****************************************************************************
+
+  async startEngine(): Promise<void> {
+    this.engineOn = true
+    await this.tools.connectClient()
+    this.addToLoop('queryBalance', ACCOUNT_POLL_MILLISECONDS).catch(() => {})
+    this.addToLoop('queryBlockheight', ACCOUNT_POLL_MILLISECONDS).catch(
+      () => {}
+    )
+    this.addToLoop('queryTransactions', TRANSACTION_POLL_MILLISECONDS).catch(
+      () => {}
+    )
+    await super.startEngine()
+  }
+
+  async killEngine(): Promise<void> {
+    await this.tools.disconnectClient()
+    await super.killEngine()
+  }
+
+  async resyncBlockchain(): Promise<void> {
+    await this.killEngine()
+    await this.clearBlockchainCache()
+    await this.startEngine()
+  }
+
+  async makeSpend(edgeSpendInfoIn: EdgeSpendInfo): Promise<EdgeTransaction> {
+    edgeSpendInfoIn = upgradeMemos(edgeSpendInfoIn, this.currencyInfo)
+    const { edgeSpendInfo, currencyCode, nativeBalance } =
+      this.makeSpendCheck(edgeSpendInfoIn)
+    const { memos = [] } = edgeSpendInfo
+    const memo: string | undefined = memos[0]?.value
+
+    if (edgeSpendInfo.spendTargets.length !== 1) {
+      throw new Error('Error: only one output allowed')
+    }
+    const { nativeAmount, publicAddress } = edgeSpendInfo.spendTargets[0]
+    if (nativeAmount == null) throw new NoAmountSpecifiedError()
+    if (publicAddress == null)
+      throw new Error('makeSpend Missing publicAddress')
+
+    const networkFee = this.networkInfo.defaultTransactionFee.amount
+    const totalNativeAmount = add(nativeAmount, networkFee)
+    if (gt(totalNativeAmount, nativeBalance)) {
+      throw new InsufficientFundsError()
+    }
+
+    // Encode a send message.
+    const msg = this.tools.methods.transfer({
+      amount: nativeAmount,
+      fromAddress: this.walletInfo.keys.bech32Address,
+      toAddress: publicAddress
+    })
+    const body = TxBody.fromPartial({ messages: [msg], memo })
+    const bodyBytes = TxBody.encode(body).finish()
+    const unsignedTxRaw = TxRaw.fromPartial({
+      bodyBytes
+    })
+    const unsignedTxHex = base16.stringify(TxRaw.encode(unsignedTxRaw).finish())
+
+    const otherParams = {
+      unsignedTxHex
+    }
+
+    const edgeTransaction: EdgeTransaction = {
+      blockHeight: 0,
+      currencyCode,
+      date: 0,
+      isSend: true,
+      memos,
+      nativeAmount: `-${totalNativeAmount}`,
+      networkFee,
+      otherParams,
+      ourReceiveAddresses: [],
+      signedTx: '',
+      txid: '',
+      walletId: this.walletId
+    }
+
+    return edgeTransaction
+  }
+
+  async signTx(
+    edgeTransaction: EdgeTransaction,
+    privateKeys: JsonObject
+  ): Promise<EdgeTransaction> {
+    const { unsignedTxHex } = edgeTransaction.otherParams ?? {}
+    if (unsignedTxHex == null) throw new Error('Missing unsignedTxHex')
+    const keys = asCosmosPrivateKeys(this.currencyInfo.pluginId)(privateKeys)
+    const txRawBytes = base16.parse(unsignedTxHex)
+    const { bodyBytes } = TxRaw.decode(txRawBytes)
+
+    const senderPubkeyBytes = base16.parse(this.walletInfo.keys.publicKey)
+    const senderPubkey = encodeSecp256k1Pubkey(senderPubkeyBytes)
+    const authInfoBytes = makeAuthInfoBytes(
+      [{ pubkey: encodePubkey(senderPubkey), sequence: this.sequence }],
+      [], // fee, but for thorchain the fee doesn't need to be defined and is automatically pulled from account
+      0, // gasLimit
+      undefined, // feeGranter
+      undefined, // feePayer (defaults to first signer)
+      1 // signMode
+    )
+
+    const signDoc = SignDoc.fromPartial({
+      accountNumber: this.accountNumber,
+      authInfoBytes,
+      bodyBytes,
+      chainId: this.networkInfo.chainId
+    })
+    const signer = await this.tools.createSigner(keys.mnemonic)
+    const signResponse = await signer.signDirect(
+      this.walletInfo.keys.bech32Address,
+      signDoc
+    )
+    const decodedSignature = decodeSignature(signResponse.signature)
+    const signedTxRaw = TxRaw.fromPartial({
+      authInfoBytes,
+      bodyBytes,
+      signatures: [decodedSignature.signature]
+    })
+    const signedTxBytes = TxRaw.encode(signedTxRaw).finish()
+    const signedTxHex = base16.stringify(signedTxBytes)
+    edgeTransaction.signedTx = signedTxHex
+    return edgeTransaction
+  }
+
+  async broadcastTx(
+    edgeTransaction: EdgeTransaction
+  ): Promise<EdgeTransaction> {
+    try {
+      const signedTxBytes = base16.parse(edgeTransaction.signedTx)
+      const client = await this.getStargateClient()
+      const txid = await client.broadcastTxSync(signedTxBytes)
+      this.warn(`SUCCESS broadcastTx\n${cleanTxLogs(edgeTransaction)}`)
+
+      edgeTransaction.txid = txid
+      edgeTransaction.date = Date.now() / 1000
+      return edgeTransaction
+    } catch (e: any) {
+      this.warn('FAILURE broadcastTx failed: ', e)
+      throw e
+    }
+  }
+
+  async getFreshAddress(_options: any): Promise<EdgeFreshAddress> {
+    const { bech32Address } = this.walletInfo.keys
+
+    return {
+      publicAddress: bech32Address
+    }
+  }
+}
+
+export async function makeCurrencyEngine(
+  env: PluginEnvironment<CosmosNetworkInfo>,
+  tools: CosmosTools,
+  walletInfo: EdgeWalletInfo,
+  opts: EdgeCurrencyEngineOptions
+): Promise<EdgeCurrencyEngine> {
+  const safeWalletInfo = asSafeCosmosWalletInfo(walletInfo)
+  const engine = new CosmosEngine(env, tools, safeWalletInfo, opts)
+
+  await engine.loadEngine()
+
+  return engine
+}