+++ src/cosmos/cosmosInfos.ts
@@ -1,0 +1,5 @@
+import { thorchainrune } from './info/thorchainruneInfo'
+
+export const cosmosPlugins = {
+  thorchainrune
+}