+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- added: Add THORChain (RUNE)
+
 ## 2.10.0 (2023-11-07)
 
 - added: Add a `yarn cli` command for debugging and testing plugins from the command line.